working compilation, bugs with roundoff
theoretically working implementation. Naive O(n^2) checks for collisions. round off error so timestep is too small. need further testing and fixes

+++ frame.h
@@ -5,14 +5,14 @@
 #include "particle.h"
 
 struct Frame {
-    float dt; // Set to inverse of desired frame rate
+    double dt; // Set to inverse of desired frame rate
     int number_of_particles;
     std::vector<Particle> all_particles;
 
     //TO DO: make an initializer for ease of testing  
 };
 
-void update_particle(Particle& p, float dt);
+void update_particle(Particle& p, double dt);
 
 void update_frame(Frame& frame);
