+++ particle.h
@@ -8,12 +8,12 @@ struct Particle {
     Vector velocity = Vector{0, 0};
     Vector acceleration = Vector{0, -9.81};
 
-    float radius = 1;
+    double radius = 1;
 };
 
 void detect_collision(Particle& p1, Particle& p2);
 
-void resolve_collision(Particle& p1, Particle& p2, float e = 1);
+void resolve_collision(Particle& p1, Particle& p2, double e = 1);
 
 void resolve_intersection(Particle& p1, Particle& p2);
